refactor(proxmox): extract route-inline components to components/proxmox/
IntervalToggle and UpdateIndicator were defined inline in routes/proxmox.tsx.
Each is now in its own file under components/proxmox/.
