+++ src/components/proxmox/ProxmoxUpdateIndicator.tsx
@@ -1,0 +1,57 @@
+import { useEffect, useRef } from 'react'
+import { Tooltip } from '@mui/material'
+import { useAtomValue } from 'jotai'
+import { proxmoxLastUpdateAtom } from '@/hooks/settingsAtom'
+
+export function UpdateIndicator({ expectedInterval }: { expectedInterval: number }) {
+  const lastUpdate = useAtomValue(proxmoxLastUpdateAtom)
+  const dotRef = useRef<HTMLDivElement>(null)
+  const pingRef = useRef<HTMLDivElement>(null)
+
+  useEffect(() => {
+    if (lastUpdate === 0) return
+
+    const dot = dotRef.current
+    const ping = pingRef.current
+
+    // Reset to active state
+    if (dot) { dot.className = 'absolute w-2 h-2 rounded-full transition-all duration-300 bg-green-500 opacity-100' }
+    if (ping) { ping.className = 'absolute w-2 h-2 bg-green-500 rounded-full animate-ping opacity-75' }
+
+    const pulseTimer = setTimeout(() => {
+      if (ping) { ping.className = 'absolute w-2 h-2 bg-green-500 rounded-full opacity-0' }
+    }, 1000)
+
+    // After twice the expected interval plus a 5s grace period, dim the dot to signal stale data
+    const lateThreshold = expectedInterval * 2 + 5000
+    const lateCheckTimer = setTimeout(() => {
+      if (dot) { dot.className = 'absolute w-2 h-2 rounded-full transition-all duration-300 bg-orange-500 opacity-30' }
+      if (ping) { ping.className = 'absolute w-2 h-2 bg-green-500 rounded-full opacity-0' }
+    }, lateThreshold)
+
+    return () => {
+      clearTimeout(pulseTimer)
+      clearTimeout(lateCheckTimer)
+    }
+  }, [lastUpdate, expectedInterval])
+
+  const lastUpdatedDate = lastUpdate > 0 ? new Date(lastUpdate) : null
+  const tooltipTitle = lastUpdatedDate
+    ? `Last updated: ${lastUpdatedDate.toLocaleTimeString()}`
+    : 'No data yet'
+
+  return (
+    <Tooltip title={tooltipTitle} placement="bottom">
+      <div className="relative inline-flex items-center justify-center w-2 h-2">
+        <div
+          ref={dotRef}
+          className="absolute w-2 h-2 rounded-full transition-all duration-300 bg-green-500 opacity-100"
+        />
+        <div
+          ref={pingRef}
+          className="absolute w-2 h-2 bg-green-500 rounded-full opacity-0"
+        />
+      </div>
+    </Tooltip>
+  )
+}
